refactor(aio): rename `CustomMdIconRegistry` to `CustomIconRegistry`
The change of Angular Material version means that the `md` prefix is
no longer appropriate.
