+++ aio/src/app/shared/custom-icon-registry.spec.ts
@@ -1,7 +1,7 @@
 import { MatIconRegistry } from '@angular/material';
-import { CustomMdIconRegistry, SvgIconInfo } from './custom-md-icon-registry';
+import { CustomIconRegistry, SvgIconInfo } from './custom-icon-registry';
 
-describe('CustomMdIconRegistry', () => {
+describe('CustomIconRegistry', () => {
   it('should get the SVG element for a preloaded icon from the cache', () => {
     const mockHttp: any = {};
     const mockSanitizer: any = {};
@@ -10,7 +10,7 @@ describe('CustomMdIconRegistry', () => {
     const svgIcons: SvgIconInfo[] = [
       { name: 'test_icon', svgSource: svgSrc }
     ];
-    const registry = new CustomMdIconRegistry(mockHttp, mockSanitizer, svgIcons);
+    const registry = new CustomIconRegistry(mockHttp, mockSanitizer, svgIcons);
     let svgElement: SVGElement;
     registry.getNamedSvgIcon('test_icon').subscribe(el => svgElement = el);
     expect(svgElement).toEqual(createSvg(svgSrc));
@@ -26,7 +26,7 @@ describe('CustomMdIconRegistry', () => {
     ];
     spyOn(MatIconRegistry.prototype, 'getNamedSvgIcon');
 
-    const registry = new CustomMdIconRegistry(mockHttp, mockSanitizer, svgIcons);
+    const registry = new CustomIconRegistry(mockHttp, mockSanitizer, svgIcons);
 
     registry.getNamedSvgIcon('other_icon');
     expect(MatIconRegistry.prototype.getNamedSvgIcon).toHaveBeenCalledWith('other_icon', undefined);
